Extract interfaces from types.ts
