+++ src/commands/game_commands/recentlyadded.ts
@@ -1,6 +1,5 @@
 import { EmbedOptions } from "eris";
 import BaseCommand, { CommandArgs, Help } from "../interfaces/base_command";
-import { QueriedSong } from "../../types";
 import dbContext from "../../database_context";
 import {
     getDebugLogHeader,
@@ -21,6 +20,7 @@ import { KmqImages } from "../../constants";
 import { IPCLogger } from "../../logger";
 import MessageContext from "../../structures/message_context";
 import { state } from "../../kmq_worker";
+import QueriedSong from "../../interfaces/queried_song";
 
 const logger = new IPCLogger("recentlyadded");
