+++ src/interfaces/embed_payload.ts
@@ -1,0 +1,17 @@
+import Eris from "eris";
+
+export default interface EmbedPayload {
+    title: string;
+    url?: string;
+    description?: string;
+    footerText?: string;
+    thumbnailUrl?: string;
+    timestamp?: Date;
+    fields?: Eris.EmbedField[];
+    author?: {
+        username: string;
+        avatarUrl: string;
+    };
+    color?: number;
+    components?: Eris.ActionRow[];
+}
