+++ src/interfaces/game_info_message.ts
@@ -1,0 +1,5 @@
+export default interface GameInfoMessage {
+    title: string;
+    message: string;
+    weight: number;
+}
