+++ src/interfaces/matched_artist.ts
@@ -1,0 +1,4 @@
+export default interface MatchedArtist {
+    id: number;
+    name: string;
+}
