+++ src/interfaces/parsed_message.ts
@@ -1,0 +1,6 @@
+export default interface ParsedMessage {
+    action: string;
+    argument: string;
+    message: string;
+    components: Array<string>;
+}
