+++ src/interfaces/player_round_result.ts
@@ -1,0 +1,8 @@
+import KmqMember from "../structures/kmq_member";
+
+export default interface PlayerRoundResult {
+    player: KmqMember;
+    streak: number;
+    expGain: number;
+    pointsEarned: number;
+}
