+++ src/interfaces/precheck_args.ts
@@ -1,0 +1,8 @@
+import Session from "../structures/session";
+import { GuildTextableMessage } from "../types";
+
+export default interface PrecheckArgs {
+    message: GuildTextableMessage;
+    session: Session;
+    errorMessage?: string;
+}
