+++ src/scripts/get-unclean-song-names.ts
@@ -1,5 +1,5 @@
-import { QueriedSong } from "../types";
 import dbContext from "../database_context";
+import QueriedSong from "../interfaces/queried_song";
 import { IPCLogger } from "../logger";
 
 const logger = new IPCLogger("get-unclean-song-names");
