+++ src/structures/music_round.ts
@@ -1,6 +1,7 @@
 import Eris from "eris";
 import { EMBED_SUCCESS_BONUS_COLOR } from "../helpers/discord_utils";
-import { PlayerRoundResult, QueriedSong } from "../types";
+import PlayerRoundResult from "../interfaces/player_round_result";
+import QueriedSong from "../interfaces/queried_song";
 import MessageContext from "./message_context";
 import Round from "./round";
 import { UniqueSongCounter } from "./song_selector";
