+++ src/structures/round.ts
@@ -1,9 +1,10 @@
 import Eris from "eris";
-import { PlayerRoundResult, QueriedSong } from "../types";
 import { state } from "../kmq_worker";
 import { UniqueSongCounter } from "./song_selector";
 import MessageContext from "./message_context";
 import { codeLine, friendlyFormattedNumber } from "../helpers/utils";
+import QueriedSong from "../interfaces/queried_song";
+import PlayerRoundResult from "../interfaces/player_round_result";
 
 export const MAX_RUNNERS_UP = 30;
 export const MAX_SCOREBOARD_PLAYERS = 30;
